[CCUBE-1199][BL] Refactoring of styling and types
